Scapin.Hooke — Type

Isotropic, linear elastic material.

Hooke{T, DIM}(μ::T, ν::T)

Create a new instance with shear modulus μ and Poisson ratio ν.

Material stability

Material stability requires that μ > 0 and -1 < ν < 1/2; these conditions are not enforced here. In other words, unstable materials can be defined.

Plane stresses vs. plane strains

In the current implementation, DIM = 2 refers to plane strain elasticity. For plane stresses, the true Poisson ratio ν should be replaced with the fictitious ratio ν̃ = ν / (1 + ν).

Scapin.avg_gradient_operator — Method

avg_gradient_operator(h)

Return the cell average of the gradient operator.

Scapin.avg_strain_displacement_operator — Method

avg_strain_displacement_operator(h)

Return the cell average of the strain-displacement operator.

Scapin.bulk_modulus — Method

bulk_modulus(C::Hooke)

Return the bulk modulus κ for the specified Hooke material.

For plane strain elasticity

κ = μ / (1 - 2ν),

and, for 3D elasticity

κ = 2/3 μ (1 + ν) / (1 - 2ν).

Scapin.element_nodes — Method

element_nodes(e, N)

Return the nodes of element e within the grid of size N.

The element is specified as a d-dimensional multi-index, e. The function returns an array of 2^d multi-indices. The size of the grid is specified by the array N (size(N, 1) = d).

Note that periodic boundary conditions are accounted for.

Scapin.element_nodes — Method

element_nodes(d)

Return the multi-indices of the d-dimensional brick element.

This function returns an object 𝔑 of type CartesianIndices. An element n ∈ 𝔑 represents the node with coordinates (x[1], …, x[d])

x[i] = (-1)^n[i] * h[i] / 2,    i = 1, …, d

See “Geometry of the reference brick element” in the docs.

Scapin.global_stiffness_operator — Method

global_stiffness_matrix(N, h, μ, ν)

Return the global stiffness matrix for periodic, homogeneous elasticity.

N: grid size

h: cell size

μ: shear modulus

ν: Poisson ratio

Scapin.gradient_operator — Method

gradient_operator(x, h)

Return the gradient operator at the specified point.

This function returns a (d+1) dimensional array D of size (d, 2, 2, …). If n is the multi-index of the node, and i is the index of a component, then D[i, n] is the partial derivative of N[n] w.r.t. x[i], evaluated at x.

h is the size of the brick element.

See “Geometry of the reference brick element”.

Scapin.integrate — Method

integrate(f, h)

Return the N-dimensional integral of f over (0, h[1]) × (0, h[2]) × … × (0, h[N]).

Uses 2-point Gauss-Legendre integration (tensorized over the N dimensions). f must take a 1-dimensional array of size N as unique input. If avg is true, the function returns the N-dimensional average.

Scapin.modal_stiffness! — Method

modal_stiffness!(K, k, N, h, C)

Compute the modal stiffness matrix K in place, for the spatial frequency k. The grid is defined by its size, N and spacing, h; C is the material.

Scapin.modal_stiffness — Method

modal_stiffness!(K, k, N, h, C)

Compute the modal stiffness matrix K, for the spatial frequency k. The grid is defined by its size, N and spacing, h; C is the material.

Scapin.modal_strain_displacement! — Method

modal_strain_displacement!(B, k, N, h)

Compute the modal strain-displacement vector B in place, for the spatial frequency k. The grid is defined by its size, N and spacing, h.

Scapin.modal_strain_displacement — Method

modal_strain_displacement(k, N, h)

Compute the modal strain-displacement vector B for the spatial frequency k. The grid is defined by its size, N and spacing, h.

Scapin.shape — Method

shape(x, h)

Return the value of the shape functions of the element, at the specified point.

This function returns a d-dimensional array N, such that N[n] is the shape function associated with node n, evaluated at x. In particular, N[n] evaluated at node m is δ[m, n] (Kronecker).

See “Shape functions” in the docs.

Scapin.strain_displacement_operator — Method

strain_displacement_operator(x, h)

Return the strain-displacement operator for the d-dimensional brick element of size h, evaluated at point x.

This function returns a (d+3) dimensional array B of size (d, d, 2, …, 2, d). If n is the multi-index of the node, and i, j, k are component indices then, the interpolated (i, j) component of the strain at x reads

ε[i, j] = Σₙ Σₖ B[i, j, n, k] * u[n, k].

See “Gradient and strain-displacement operators”.