Scapin.Hooke — Type

Isotropic, linear elastic material.

Hooke{T, DIM}(μ::T, ν::T)

Create a new instance with shear modulus μ and Poisson ratio ν.

Material stability

Material stability requires that μ > 0 and -1 < ν < 1/2; these conditions are not enforced here. In other words, unstable materials can be defined.

Plane stresses vs. plane strains

In the current implementation, DIM = 2 refers to plane strain elasticity. For plane stresses, the true Poisson ratio ν should be replaced with the fictitious ratio ν̃ = ν / (1 + ν).

Scapin.avg_strain_displacement_matrix — Method

avg_strain_displacement_matrix(h)

Return the strain-displacement matrix, averaged over the whole element.

Scapin.bulk_modulus — Method

bulk_modulus(C::Hooke)

Return the bulk modulus κ for the specified Hooke material.

For plane strain elasticity

κ = μ / (1 - 2ν),

and, for 3D elasticity

κ = 2/3 μ (1 + ν) / (1 - 2ν).

Scapin.cell_vertices — Method

cell_vertices(cell, linear)

Return the vertices of the cell (specified as a multi-index) as an array of indices. The grid size is defined by linear (an instance of LinearIndices).

Scapin.integrate — Method

integrate(f, h)

Return the N-dimensional integral of f over (0, h[1]) × (0, h[2]) × … × (0, h[N]).

Uses 2-point Gauss-Legendre integration (tensorized over the N dimensions). f must take a 1-dimensional array of size N as unique input. If avg is true, the function returns the N-dimensional average.

Scapin.modal_stiffness! — Method

modal_stiffness!(K, k, N, h, C)

Compute the modal stiffness matrix K in place, for the spatial frequency k. The grid is defined by its size, N and spacing, h; C is the material.

Scapin.modal_stiffness — Method

modal_stiffness!(K, k, N, h, C)

Compute the modal stiffness matrix K, for the spatial frequency k. The grid is defined by its size, N and spacing, h; C is the material.

Scapin.modal_strain_displacement! — Method

modal_strain_displacement!(B, k, N, h)

Compute the modal strain-displacement vector B in place, for the spatial frequency k. The grid is defined by its size, N and spacing, h.

Scapin.modal_strain_displacement — Method

modal_strain_displacement(k, N, h)

Compute the modal strain-displacement vector B for the spatial frequency k. The grid is defined by its size, N and spacing, h.

Scapin.shape — Method

shape(ξ, k)

Return the value of the shape functions or their derivatives, at the specified point.

ξ is the d-dimensional array of the reduced coordinates (d: number of spatial dimensions), such that 0 ≤ ξ[i] ≤ 1.

Vertices of the element are referred to with their cartesian index n, a d-dimensional array such that n[1], …, n[d] ∈ {1, 2}. The reduced coordinates of vertex n are n .- 1.

For k == 0, the function returns a d-dimensional array N, such that N[n] is the shape function associated with vertex n. In other words,

size(N, 1) = … = size(N, d) = 2

.

shape(n .- 1, 0)[m] = 0    (m ≠ n),
shape(n .- 1, 0)[n] = 1.

For k == 1, …, d, the function returns the array of the derivatives of the shape functions at ξ, with respect to ξ[k].

Scapin.stiffness_matrix — Method

global_stiffness_matrix(N, h, μ, ν)

Return the global stiffness matrix for periodic, homogeneous elasticity.

N: grid size

h: cell size

μ: shear modulus

ν: Poisson ratio

Scapin.strain_displacement_matrix — Method

strain_displacement_operator(x, h)

Return the strain-displacement matrix for the d-dimensional element of size h.

The strain-displacement operator maps the array of nodal displacements u onto the array of strains at x. u is a d+1 dimensional array (d: number of spatial dimensions) such that

size(u, 1) = … = size(u, d) = 2``,
size(u, d+1) = d

and the i-th component of the displacement of vertex n is u[n..., i]. See shape_function for the numbering of vertices. The strain at x, ε is a 2-dimensional array array such that

size(ε, 1) = size(ε, 2) = d

and

ε[i, j] = Σₙ Σₖ B[i, j, n..., k] * u[n..., k],

where B is the strain-displacement operator B returned by the present function. This d+3-dimensional array is such that

size(B, 1) = size(B, 2) = d,
size(B, 3) = … = size(B, d+2) = 2,
size(B, d+3) = d.