API of the Scapin library

Scapin.Bri17

Scapin.Bri17.modal_stiffness! — Method

modal_stiffness!(K, n, N, h, C)

Compute the modal stiffness matrix K in place, for the spatial frequency n.

The grid is defined by its size, N and spacing, h. The spatial frequency is defined by n ∈ CartesianIndices(1:N[1], …, 1:N[d]). The (homogeneous) constitutive material is specified by C.

Scaling of the modal stiffness matrix

The modal stiffness matrix introduced above differs from the matrix initially introduced in Ref. DOI:10.1002/nme.5263 by a factor h[1] * … * h[d]. More precisely,

K̂{Scapin} = h[1] * … * h[d] * K̂{10.1002/nme.5263}

Such rescaling makes the relation between modal and nodal stiffness operators more natural (the former is the DFT of the latter).

Scapin.Bri17.modal_stiffness — Method

modal_stiffness(n, N, h, C)

Return the modal stiffness matrix K.

See modal_stiffness! for a description of the parameters.

Scapin.Bri17.modal_strain_displacement! — Method

modal_strain_displacement!(B, n, N, h)

Compute the modal strain-displacement vector B in place, for the spatial frequency n.

The grid is defined by its size, N and spacing, h. The spatial frequency is defined by n ∈ CartesianIndices(1:N[1], …, 1:N[d]).

Scapin.Bri17.modal_strain_displacement — Method

modal_strain_displacement(n, N, h)

Return the modal strain-displacement vector B.

See modal_strain_displacement! for a description of the parameters.

Scapin.Brick

Scapin.Brick.avg_gradient_operator — Method

avg_gradient_operator(h)

Return the cell-average of the gradient operator.

Scapin.Brick.avg_strain_displacement_operator — Method

avg_strain_displacement_operator(h)

Return the cell average of the strain-displacement operator.

Scapin.Brick.global_stiffness_operator — Method

global_stiffness_operator(N, h, μ, ν)

Return the global stiffness operator for periodic, homogeneous elasticity.

The grid size is N, the cell size is h. The constitutive material is homogeneous, elastic linear and isotropic with stiffness C.

The global stiffness operator K is a 2d+2-dimensional array of size (d, N[1], …, N[d], d, N[1], …, N[d]), such that the strain energy of the system reads

U = u[i, p] * K[i, p, j, q] * u[j, q] / 2,

where

i, j ∈ {1, …, d}: component indices,

p, q ∈ CartesianIndices(1:N[1], …, 1:N[d]): node indices,

u[i, p]: i-th component of the displacement of node p.

Note

Assembly of the global stiffness opertor is done under the assumption of periodicity.

Scapin.Brick.global_strain_displacement_operator — Method

global_strain_displacement_operator(N, h)

Return the global strain-displacement operator for periodic boundary conditions.

The grid is defined by its size N and its spacing h.

The global strain-displacement operator B is a d+2-dimensional array of size (d, d, N[1], …, N[d]), such that the average strain within element p reads

ε[i, j, p] = B[i, j, p, k, q] * u[k, q],

where

i, j, k ∈ {1, …, d}: component indices,

p, q ∈ CartesianIndices(1:N[1], …, 1:N[d]): node indices,

ε[i, j, p]: (i, j)-th component of the average strain in element p,

u[k, q]: i-th component of the displacement of node q.

Note

Assembly of the global strain-displacement operator is done under the assumption of periodicity.

Scapin.Brick.gradient_operator — Method

gradient_operator(x, h)

Return the gradient operator at the specified point.

This function returns a (d+1) dimensional array D of size (d, 2, 2, …). If i is the index of a component and p the CartesianIndex of the node, then D[i, p] is the partial derivative of N[p] w.r.t. x[i], evaluated at x.

h is the size of the brick element.

Scapin.Brick.integrate — Method

integrate(f, h)

Return the d-dimensional integral of f over (0, h[1]) × (0, h[2]) × … × (0, h[d]).

Uses 2-point Gauss-Legendre integration (tensorized over the d dimensions). f must take a 1-dimensional array of size d as unique input. If avg is true, the function returns the N-dimensional average.

Scapin.Brick.shape — Method

shape(x, h)

Return the value of the shape functions of the element, at the specified point.

This function returns a d-dimensional array N, such that N[p] is the shape function associated with node p, evaluated at x. In particular, N[p] evaluated at node q is δ[p, q] (Kronecker).

Scapin.Brick.stiffness_operator — Method

stiffness_operator(h, C)

Return the stifness operator for the brick element of size h, and Hooke material C.

The stiffness operator K delivers the strain energy associated to the nodal displacements u

U = u[i, p] * K[i, p, j, q] * u[j, q] / 2,

where i, j ∈ {1, …, d} are component indices and p, q ∈ CartesianIndices(1:2, …, 1:2).

Scapin.Brick.strain_displacement_operator — Method

strain_displacement_operator(x, h)

Return the strain-displacement operator for the d-dimensional brick element of size h, evaluated at point x.

This function returns a (d+3) dimensional array B of size (d, d, 2, …, 2, d). If p is the CartesianIndex of the node, and i, j, k are component indices then, the interpolated (i, j) component of the strain at x reads

ε[i, j] = Σₖ Σₚ B[i, j, k, p] * u[k, p].

Scapin.Elasticity

Scapin.Elasticity.Hooke — Type

Isotropic, linear elastic material.

Hooke{T, DIM}(μ::T, ν::T)

Create a new instance with shear modulus μ and Poisson ratio ν.

Material stability

Material stability requires that μ > 0 and -1 < ν < 1/2; these conditions are not enforced here. In other words, unstable materials can be defined.

Plane stresses vs. plane strains

In the current implementation, DIM = 2 refers to plane strain elasticity. For plane stresses, the true Poisson ratio ν should be replaced with the fictitious ratio ν̃ = ν / (1 + ν).

Scapin.Elasticity.bulk_modulus — Method

bulk_modulus(C::Hooke)

Return the bulk modulus κ for the specified Hooke material.

For plane strain elasticity

κ = μ / (1 - 2ν),

and, for 3D elasticity

κ = 2/3 μ (1 + ν) / (1 - 2ν).

Scapin.Grid

Scapin.Grid.cell_vertices — Method

cell_vertices(d)

Return the local multi-indices of vertices of the d-dimensional brick element.

This function returns an object 𝓛 of type CartesianIndices. An element v ∈ 𝓛 represents the vertex with coordinates (x[1], …, x[d])

x[i] = (-1)^v[i] * h[i] / 2,    i = 1, …, d

See Geometry of the reference brick element.

Scapin.Grid.cell_vertices — Method

cell_vertices(p, N)

Return the global multi-indices of the vertices of a grid-cell.

The cell is specified through its CartesianIndex, p; the size of the grid is defined by N.

This function returns a d-dimensional array 𝒢 of CartesianIndex, such that 𝒢[l] is the global index of the vertex with local index l ∈ CartesianIndices(1:2, …, 1:2).

Note that periodic boundary conditions are used.